+++ src/viagem/viagem.module.ts
@@ -2,7 +2,7 @@ import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { Viagem } from "./entities/viagem.entity";
 import { UsuarioModule } from "../usuario/usuario.module";
-import { VeiculoModule } from "../veiculos/veiculo.module";
+import { VeiculoModule } from "../veiculo/veiculo.module";
 
 @Module ({
     imports: [TypeOrmModule.forFeature([Viagem]), VeiculoModule, UsuarioModule],
